Fix clear complete bug
